Acrescentei o GlobalState
