+++ pokedex/src/App.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { Div } from './style';
 import Router from './routes/Router'
+import GlobalState from './global/GlobalState';
 
 function App() {
   return (
-    <Router>
-       
-    </Router>
+    <GlobalState>
+      <Router />
+    </GlobalState>
     
   )
 }
