+++ pokedex/src/pages/PokedexPage/PokedexPage.js
@@ -1,5 +1,9 @@
 import React from 'react'
 
-export default function PokedexPage() {
-  return <div>PokedexPage</div>
+const PokedexPage = () => {
+  return (
+    <div>PokedexPage</div>
+  )
 }
+
+export default PokedexPage
